This removes some sources of warnings, and removes all "unused variable" warnings.
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@937 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 48cd2cde96a63e7d55dd13972db54e8b50b06f9c

+++ io/mesh_output/vtk/PdVTK.cxx
@@ -264,7 +264,7 @@ void expandRingPostProcess(double current_time, vtkSmartPointer<vtkUnstructuredG
 	 * Calculate kinetic energy
 	 * HERE WE LOOP OVER ALL OWNED POINTS on processor
 	 */
-	for(size_t n=0;n<grid->GetNumberOfPoints();n++){
+	for(int n=0;n<grid->GetNumberOfPoints();n++){
 		vtkIdType localId = n;
 		double *v = vData->GetTuple3(localId);
 		double volume = volData->GetTuple1(localId);
